+++ packages/at_persistence_secondary_server/test/hive_keystore_impl_test.dart
@@ -4,7 +4,6 @@ import 'dart:io';
 
 import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
-import 'package:at_persistence_secondary_server/src/impl/hive/hive_instances.dart';
 import 'package:at_persistence_secondary_server/hive.dart';
 import 'package:crypto/crypto.dart';
 import 'package:test/test.dart';